+++ src/context/userContext.ts
@@ -1,0 +1,4 @@
+import { createContext } from 'react';
+
+type TUserContext = { user: User | null; setUser: SetState<User | null> };
+export const UserContext = createContext({} as TUserContext);
